pyproject.toml
[tool.poetry]
name = "Quart-Rest-Template"
version = "0.1.0"
description = ""
authors = ["Sebastian Henao <>"]
package-mode = false

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.dependencies]
python = "^3.12"
quart = "^0.20.0"
python-dotenv = "^1.2.1"
quart-jwt-extended = "^0.1.0"
asyncpg = "^0.30.0"
quart-schema = "^0.22.0"
tortoise-orm = "^0.25.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
black = "25.1.0"